backend/requirements.txt
fastapi
uvicorn
python-multipart
filetype
google-genai
requests
beautifulsoup4
python-dotenv
Pillow
opencv-python
pytesseract
pdfplumber
exifread
numpy
